rename files; add obtain_external_data_files.py; refresh complete data
