+++ src/clincoded/tests/data/insertsComplete/readme.md
@@ -2,5 +2,5 @@ The *.json files in this directory represent the 'complete'/non-test data for th
 
 The scripts used to generate the *.json files are in the scripts subdirectory
 
-**gene.json** - 39,292 entries - generated 2015/07/24 using data provided by kgliu0101 (~1 month old?)
-**orphaPhenotype.json** - 9,123 entries - generated 2015/07/24 using data provided by kgliu0101 (~1 month old?)
+**gene.json** - 39,446 entries - data grabbed 2015/07/24
+**orphaPhenotype.json** - 9,193 entries - data grabbed 2015/07/24 (1.1.4 / 4.1.3 ORPHANET [2014-10-03])
