+++ src/clincoded/tests/data/insertsComplete/scripts/rawData/readme.md
@@ -1,0 +1,1 @@
+Output directory for the files generated by obtain_external_data_files.py
